Extract the base backend test into a class stored in the main service discovery artifact
So, it avoids code duplication and make sure all backends pass the tests.
